+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Should prevent random crash [#172]
+
 ## [1.3.35] - 2019-05-19
 ### Added
 - Config to add fixed values or multipliers to specific dimensions (Dimension Value Factor under difficulty category) [#154]